// sbah1-main/types.ts
export type Language = 'ar' | 'fr' | 'en';
export type TranslatableString = { [key in Language]?: string; };
// ...
export type ApprovalStatus = 'Pending' | 'Approved' | 'Rejected';

export interface RestaurantManagementData {
  id: string;
  name: TranslatableString | string;
  ownerName: string;
  ownerEmail: string;
  ownerPhone: string;
  joinDate: string;
  approvalStatus: ApprovalStatus;
  isActive: boolean;
  location?: { latitude: number; longitude: number; addressText: string };
  coverPhotoUrl?: string;
  cuisine?: TranslatableString | string;
}

export interface AiAnalysis {
  summary: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
}